Versão 1.0.0

+++ src/app/modules/product/components/product-dashboard/product-dashboard.component.ts
@@ -67,7 +67,7 @@ export class ProductDashboardComponent implements OnInit {
   }
 
   cancel() {
-    if (!this.editing() && this.product.file) {
+    if (!this.editing() && this.product.file && this.product.file.id) {
       this.imageService.destroyImage(this.product.file.id)
         .subscribe();
     }
